pyproject.toml
[tool.poetry]
name = "github-bot"
version = "0.1.0"
description = ""
authors = ["Martin More <martinmore@gmail.com>"]
readme = "README.md"
packages = [{include = "github_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["debug-server"], version = "^0.151.2"}
fastapi = "^0.88.0"
httpx = "^0.23.1"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.2"
black = "^23.3.0"
isort = "^5.12.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
